+++ activisionpro.h
@@ -14,7 +14,7 @@
 // Public API:
 //   struct activisionpro_state *activisionpro_init(void *data, uint32_t len, int32_t sample_rate);
 //   void activisionpro_free(struct activisionpro_state *s);
-//   void activisionpro_get_audio(struct activisionpro_state *s, int16_t *output, int32_t frames);
+//   void activisionpro_get_audio(struct activisionpro_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -2106,7 +2106,7 @@ static void activisionpro_free(struct activisionpro_state *s) {
 }
 
 // [=]===^=[ activisionpro_get_audio ]============================================================[=]
-static void activisionpro_get_audio(struct activisionpro_state *s, int16_t *output, int32_t frames) {
+static void activisionpro_get_audio(struct activisionpro_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -2134,7 +2134,7 @@ static void activisionpro_api_free(void *state) {
 }
 
 // [=]===^=[ activisionpro_api_get_audio ]========================================================[=]
-static void activisionpro_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void activisionpro_api_get_audio(void *state, float *output, int32_t frames) {
 	activisionpro_get_audio((struct activisionpro_state *)state, output, frames);
 }
