+++ bendaglish.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct bendaglish_state *bendaglish_init(void *data, uint32_t len, int32_t sample_rate);
 //   void bendaglish_free(struct bendaglish_state *s);
-//   void bendaglish_get_audio(struct bendaglish_state *s, int16_t *output, int32_t frames);
+//   void bendaglish_get_audio(struct bendaglish_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1866,7 +1866,7 @@ static void bendaglish_free(struct bendaglish_state *s) {
 }
 
 // [=]===^=[ bendaglish_get_audio ]===============================================================[=]
-static void bendaglish_get_audio(struct bendaglish_state *s, int16_t *output, int32_t frames) {
+static void bendaglish_get_audio(struct bendaglish_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1894,7 +1894,7 @@ static void bendaglish_api_free(void *state) {
 }
 
 // [=]===^=[ bendaglish_api_get_audio ]===========================================================[=]
-static void bendaglish_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void bendaglish_api_get_audio(void *state, float *output, int32_t frames) {
 	bendaglish_get_audio((struct bendaglish_state *)state, output, frames);
 }
