+++ digibooster.h
@@ -8,7 +8,7 @@
 // Public API:
 //   struct digibooster_state *digibooster_init(void *data, uint32_t len, int32_t sample_rate);
 //   void digibooster_free(struct digibooster_state *s);
-//   void digibooster_get_audio(struct digibooster_state *s, int16_t *output, int32_t frames);
+//   void digibooster_get_audio(struct digibooster_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1366,7 +1366,7 @@ static void digibooster_free(struct digibooster_state *s) {
 }
 
 // [=]===^=[ digibooster_get_audio ]==============================================================[=]
-static void digibooster_get_audio(struct digibooster_state *s, int16_t *output, int32_t frames) {
+static void digibooster_get_audio(struct digibooster_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1394,7 +1394,7 @@ static void digibooster_api_free(void *state) {
 }
 
 // [=]===^=[ digibooster_api_get_audio ]==========================================================[=]
-static void digibooster_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void digibooster_api_get_audio(void *state, float *output, int32_t frames) {
 	digibooster_get_audio((struct digibooster_state *)state, output, frames);
 }
