+++ digitalmugician.h
@@ -9,7 +9,7 @@
 // Public API:
 //   struct digitalmugician_state *digitalmugician_init(void *data, uint32_t len, int32_t sample_rate);
 //   void digitalmugician_free(struct digitalmugician_state *s);
-//   void digitalmugician_get_audio(struct digitalmugician_state *s, int16_t *output, int32_t frames);
+//   void digitalmugician_get_audio(struct digitalmugician_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1192,7 +1192,7 @@ static void digitalmugician_free(struct digitalmugician_state *s) {
 }
 
 // [=]===^=[ digitalmugician_get_audio ]==========================================================[=]
-static void digitalmugician_get_audio(struct digitalmugician_state *s, int16_t *output, int32_t frames) {
+static void digitalmugician_get_audio(struct digitalmugician_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1220,7 +1220,7 @@ static void digitalmugician_api_free(void *state) {
 }
 
 // [=]===^=[ digitalmugician_api_get_audio ]======================================================[=]
-static void digitalmugician_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void digitalmugician_api_get_audio(void *state, float *output, int32_t frames) {
 	digitalmugician_get_audio((struct digitalmugician_state *)state, output, frames);
 }
