+++ facethemusic.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct facethemusic_state *facethemusic_init(void *data, uint32_t len, int32_t sample_rate);
 //   void facethemusic_free(struct facethemusic_state *s);
-//   void facethemusic_get_audio(struct facethemusic_state *s, int16_t *output, int32_t frames);
+//   void facethemusic_get_audio(struct facethemusic_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -2031,7 +2031,7 @@ static void facethemusic_free(struct facethemusic_state *s) {
 }
 
 // [=]===^=[ facethemusic_get_audio ]=============================================================[=]
-static void facethemusic_get_audio(struct facethemusic_state *s, int16_t *output, int32_t frames) {
+static void facethemusic_get_audio(struct facethemusic_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -2059,7 +2059,7 @@ static void facethemusic_api_free(void *state) {
 }
 
 // [=]===^=[ facethemusic_api_get_audio ]=========================================================[=]
-static void facethemusic_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void facethemusic_api_get_audio(void *state, float *output, int32_t frames) {
 	facethemusic_get_audio((struct facethemusic_state *)state, output, frames);
 }
