+++ fashiontracker.h
@@ -44,7 +44,7 @@
 // Public API:
 //   struct fashiontracker_state *fashiontracker_init(void *data, uint32_t len, int32_t sample_rate);
 //   void fashiontracker_free(struct fashiontracker_state *s);
-//   void fashiontracker_get_audio(struct fashiontracker_state *s, int16_t *output, int32_t frames);
+//   void fashiontracker_get_audio(struct fashiontracker_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -462,7 +462,7 @@ static void fashiontracker_free(struct fashiontracker_state *s) {
 }
 
 // [=]===^=[ fashiontracker_get_audio ]===========================================================[=]
-static void fashiontracker_get_audio(struct fashiontracker_state *s, int16_t *output, int32_t frames) {
+static void fashiontracker_get_audio(struct fashiontracker_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -490,7 +490,7 @@ static void fashiontracker_api_free(void *state) {
 }
 
 // [=]===^=[ fashiontracker_api_get_audio ]=======================================================[=]
-static void fashiontracker_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void fashiontracker_api_get_audio(void *state, float *output, int32_t frames) {
 	fashiontracker_get_audio((struct fashiontracker_state *)state, output, frames);
 }
