+++ futurecomposer.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct futurecomposer_state *futurecomposer_init(void *data, uint32_t len, int32_t sample_rate);
 //   void futurecomposer_free(struct futurecomposer_state *s);
-//   void futurecomposer_get_audio(struct futurecomposer_state *s, int16_t *output, int32_t frames);
+//   void futurecomposer_get_audio(struct futurecomposer_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1271,7 +1271,7 @@ static void futurecomposer_free(struct futurecomposer_state *s) {
 }
 
 // [=]===^=[ futurecomposer_get_audio ]============================================================[=]
-static void futurecomposer_get_audio(struct futurecomposer_state *s, int16_t *output, int32_t frames) {
+static void futurecomposer_get_audio(struct futurecomposer_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1299,7 +1299,7 @@ static void futurecomposer_api_free(void *state) {
 }
 
 // [=]===^=[ futurecomposer_api_get_audio ]========================================================[=]
-static void futurecomposer_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void futurecomposer_api_get_audio(void *state, float *output, int32_t frames) {
 	futurecomposer_get_audio((struct futurecomposer_state *)state, output, frames);
 }
