+++ gamemusiccreator.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct gamemusiccreator_state *gamemusiccreator_init(void *data, uint32_t len, int32_t sample_rate);
 //   void gamemusiccreator_free(struct gamemusiccreator_state *s);
-//   void gamemusiccreator_get_audio(struct gamemusiccreator_state *s, int16_t *output, int32_t frames);
+//   void gamemusiccreator_get_audio(struct gamemusiccreator_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -547,7 +547,7 @@ static void gamemusiccreator_free(struct gamemusiccreator_state *s) {
 }
 
 // [=]===^=[ gamemusiccreator_get_audio ]=========================================================[=]
-static void gamemusiccreator_get_audio(struct gamemusiccreator_state *s, int16_t *output, int32_t frames) {
+static void gamemusiccreator_get_audio(struct gamemusiccreator_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -575,7 +575,7 @@ static void gamemusiccreator_api_free(void *state) {
 }
 
 // [=]===^=[ gamemusiccreator_api_get_audio ]=====================================================[=]
-static void gamemusiccreator_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void gamemusiccreator_api_get_audio(void *state, float *output, int32_t frames) {
 	gamemusiccreator_get_audio((struct gamemusiccreator_state *)state, output, frames);
 }
