+++ hippel.h
@@ -17,7 +17,7 @@
 // Public API:
 //   struct hippel_state *hippel_init(void *data, uint32_t len, int32_t sample_rate);
 //   void hippel_free(struct hippel_state *s);
-//   void hippel_get_audio(struct hippel_state *s, int16_t *output, int32_t frames);
+//   void hippel_get_audio(struct hippel_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -2806,7 +2806,7 @@ static void hippel_free(struct hippel_state *s) {
 }
 
 // [=]===^=[ hippel_get_audio ]===================================================================[=]
-static void hippel_get_audio(struct hippel_state *s, int16_t *output, int32_t frames) {
+static void hippel_get_audio(struct hippel_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -2834,7 +2834,7 @@ static void hippel_api_free(void *state) {
 }
 
 // [=]===^=[ hippel_api_get_audio ]===============================================================[=]
-static void hippel_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void hippel_api_get_audio(void *state, float *output, int32_t frames) {
 	hippel_get_audio((struct hippel_state *)state, output, frames);
 }
