+++ med.h
@@ -9,7 +9,7 @@
 // Public API:
 //   struct med_state *med_init(void *data, uint32_t len, int32_t sample_rate);
 //   void med_free(struct med_state *s);
-//   void med_get_audio(struct med_state *s, int16_t *output, int32_t frames);
+//   void med_get_audio(struct med_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -1105,7 +1105,7 @@ static void med_free(struct med_state *s) {
 }
 
 // [=]===^=[ med_get_audio ]======================================================================[=]
-static void med_get_audio(struct med_state *s, int16_t *output, int32_t frames) {
+static void med_get_audio(struct med_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -1133,7 +1133,7 @@ static void med_api_free(void *state) {
 }
 
 // [=]===^=[ med_api_get_audio ]==================================================================[=]
-static void med_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void med_api_get_audio(void *state, float *output, int32_t frames) {
 	med_get_audio((struct med_state *)state, output, frames);
 }
