+++ oktalyzer.h
@@ -7,7 +7,7 @@
 // Public API:
 //   struct oktalyzer_state *oktalyzer_init(void *data, uint32_t len, int32_t sample_rate);
 //   void oktalyzer_free(struct oktalyzer_state *s);
-//   void oktalyzer_get_audio(struct oktalyzer_state *s, int16_t *output, int32_t frames);
+//   void oktalyzer_get_audio(struct oktalyzer_state *s, float *output, int32_t frames);
 
 #pragma once
 
@@ -746,7 +746,7 @@ static void oktalyzer_free(struct oktalyzer_state *s) {
 }
 
 // [=]===^=[ oktalyzer_get_audio ]=================================================================[=]
-static void oktalyzer_get_audio(struct oktalyzer_state *s, int16_t *output, int32_t frames) {
+static void oktalyzer_get_audio(struct oktalyzer_state *s, float *output, int32_t frames) {
 	while(frames > 0) {
 		int32_t remain = s->paula.samples_per_tick - s->paula.tick_offset;
 		if(remain > frames) {
@@ -774,7 +774,7 @@ static void oktalyzer_api_free(void *state) {
 }
 
 // [=]===^=[ oktalyzer_api_get_audio ]=============================================================[=]
-static void oktalyzer_api_get_audio(void *state, int16_t *output, int32_t frames) {
+static void oktalyzer_api_get_audio(void *state, float *output, int32_t frames) {
 	oktalyzer_get_audio((struct oktalyzer_state *)state, output, frames);
 }
